Generalize tileset processing
Allow stopping the traversal in forEachTile,
and allow creating a new tileset object
instead of modifying the tileset in-place
