+++ specs/ImplicitTilesetProcessorSpec.ts
@@ -57,6 +57,7 @@ describe("BasicTilesetProcessor on implicit input", function () {
     await tilesetProcessor.forEachTile(async (traversedTile: TraversedTile) => {
       const contentUris = traversedTile.getFinalContents().map((c) => c.uri);
       actualContentUris.push(contentUris);
+      return true;
     });
     await tilesetProcessor.end();
